EN-4505: fix after review
