+++ cmd/node/config/config.toml
@@ -18,14 +18,14 @@
    # If this flag is set to true, the node won't delete any database between epochs
    FullArchive = false
 
-   # NumOfEpochsToKeep - if the flag above is set to false, this will set the number of epochs to keep in the storage.
+   # NumEpochsToKeep - if the flag above is set to false, this will set the number of epochs to keep in the storage.
    # Epochs older that (current epoch - NumOfEpochsToKeep) will be removed
-   NumOfEpochsToKeep = 3
+   NumEpochsToKeep = 3
 
-   # NumOfActivePersisters - this will set the number of persisters to keep active at a moment. This works for both
+   # NumActivePersisters - this will set the number of persisters to keep active at a moment. This works for both
    # full archive or not nodes. In case of a node which doesn't have a full archive, it has to be smaller or equal
    # to the NumOfEpochsToKeep flag
-   NumOfActivePersisters = 2
+   NumActivePersisters = 2
 
 [Explorer]
    Enabled = false
