+++ config/config.go
@@ -121,9 +121,9 @@ type NodeConfig struct {
 
 // StoragePruningConfig will hold settings relates to storage pruning
 type StoragePruningConfig struct {
-	FullArchive           bool
-	NumOfEpochsToKeep     uint64
-	NumOfActivePersisters uint64
+	FullArchive         bool
+	NumEpochsToKeep     uint64
+	NumActivePersisters uint64
 }
 
 // KadDhtPeerDiscoveryConfig will hold the kad-dht discovery config settings
