+++ storage/errors.go
@@ -48,3 +48,15 @@ var ErrNilEpochStartNotifier = errors.New("nil epoch start notifier")
 
 // ErrNilPersisterFactory signals that a nil persister factory has been provided
 var ErrNilPersisterFactory = errors.New("nil persister factory")
+
+// ErrDestroyingUnit signals that the destroy unit method did not manage to destroy all the persisters in a pruning storer
+var ErrDestroyingUnit = errors.New("destroy unit didn't remove all the persisters")
+
+// ErrNilConfig signals that a nil configuration has been received
+var ErrNilConfig = errors.New("nil config")
+
+// ErrNilShardCoordinator signals that a nil shard coordinator has been provided
+var ErrNilShardCoordinator = errors.New("nil shard coordinator")
+
+// ErrEmptyUniqueIdentifier signals that an empty unique ID has been provided
+var ErrEmptyUniqueIdentifier = errors.New("empty unique identifier")
