+++ storage/factory/common.go
@@ -1,0 +1,44 @@
+package factory
+
+import (
+	"path/filepath"
+
+	"github.com/ElrondNetwork/elrond-go/config"
+	"github.com/ElrondNetwork/elrond-go/storage/storageUnit"
+)
+
+// GetCacherFromConfig will return the cache config needed for storage unit from a config came from the toml file
+func GetCacherFromConfig(cfg config.CacheConfig) storageUnit.CacheConfig {
+	return storageUnit.CacheConfig{
+		Size:   cfg.Size,
+		Type:   storageUnit.CacheType(cfg.Type),
+		Shards: cfg.Shards,
+	}
+}
+
+// GetDBFromConfig will return the db config needed for storage unit from a config came from the toml file
+func GetDBFromConfig(cfg config.DBConfig, uniquePath string) storageUnit.DBConfig {
+	return storageUnit.DBConfig{
+		FilePath:          filepath.Join(uniquePath, cfg.FilePath),
+		Type:              storageUnit.DBType(cfg.Type),
+		MaxBatchSize:      cfg.MaxBatchSize,
+		BatchDelaySeconds: cfg.BatchDelaySeconds,
+		MaxOpenFiles:      cfg.MaxOpenFiles,
+	}
+}
+
+// GetBloomFromConfig will return the bloom config needed for storage unit from a config came from the toml file
+func GetBloomFromConfig(cfg config.BloomFilterConfig) storageUnit.BloomConfig {
+	var hashFuncs []storageUnit.HasherType
+	if cfg.HashFunc != nil {
+		hashFuncs = make([]storageUnit.HasherType, 0)
+		for _, hf := range cfg.HashFunc {
+			hashFuncs = append(hashFuncs, storageUnit.HasherType(hf))
+		}
+	}
+
+	return storageUnit.BloomConfig{
+		Size:     cfg.Size,
+		HashFunc: hashFuncs,
+	}
+}
