+++ storage/leveldb/leveldb.go
@@ -215,9 +215,7 @@ func (s *DB) DestroyClosed() error {
 
 	s.dbClosed <- struct{}{}
 
-	err := os.RemoveAll(s.path)
-
-	return err
+	return os.RemoveAll(s.path)
 }
 
 // IsInterfaceNil returns true if there is no value under the interface
